% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nalmfun.R
\name{age}
\alias{age}
\title{Calculate age from date of birth}
\usage{
age(dob, current = Sys.Date())
}
\arguments{
\item{dob}{: Date of birth (compulsory)}

\item{current:}{Date as which you want to calculate age at . By default its current date}
}
\value{
integer
}
\description{
This function return age of the person based on date of birth
}
\examples{
df<- c(as.Date("1996-10-10"),as.Date("1995-10-10"))
age(df,as.Date("2021-02-10"))
age(df)
}
\author{
Krishna Kumar Shrestha
}
