% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nalmfun.R
\name{date_clean}
\alias{date_clean}
\title{Change all date in standard format if it is in character}
\usage{
date_clean(date)
}
\arguments{
\item{date}{: date that needs to be changed (must be character)}
}
\value{
date
}
\description{
This function return standard format date if date is in character. if function fails to identify the date it returns the todays date
}
\examples{
df<- c("1996-10-10","12-10/1996")
date_clean(df)
}
\author{
Krishna Kumar Shrestha
}
