% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/nalmfun.R
\docType{data}
\name{nalm2009}
\alias{nalm2009}
\title{Nepal Mortality table (2009)}
\format{
A data.frame with 2 columns
\describe{
\item{age}{age of the person}
\item{qx}{mortality rate of the corresponding age}
}
}
\source{
Beema Pradhikaran
}
\usage{
nalm2009

nalm2009()
}
\value{
A data frame
}
\description{
Data

This function has been designed to return the NALM mortality table of year 2009
}
\examples{
nalm2009()
}
\author{
Krishna Kumar Shrestha
}
\keyword{datasets}
